move encode_prompt and lpw encode_prompt to base class
